Test if files exist before making the sym links

+++ ettercap/scripts/dependencies.sh
@@ -35,7 +35,9 @@ if [ "$1" = "install" ]; then
 
   fi
   
-  ln -s /usr/lib/libpcap.so /usr/lib/libpcap.so.1.3
+  if [ ! -f /usr/lib/libpcap.so ] && [ -f /usr/lib/libpcap.so.1.3 ]; then
+  	ln -s /usr/lib/libpcap.so /usr/lib/libpcap.so.1.3
+  fi
 
   touch /etc/config/ettercap
   echo "config ettercap 'module'" > /etc/config/ettercap
