+++ wps/scripts/dependencies.sh
@@ -25,7 +25,9 @@ if [ "$1" = "install" ]; then
 
   fi
   
-  ln -s /usr/lib/libpcap.so /usr/lib/libpcap.so.1.3
+  if [ ! -f /usr/lib/libpcap.so ] && [ -f /usr/lib/libpcap.so.1.3 ]; then
+  	ln -s /usr/lib/libpcap.so /usr/lib/libpcap.so.1.3
+  fi
 
   touch /etc/config/wps
   echo "config wps 'module'" > /etc/config/wps
